smdx-qbserver
**SCRIPT: smdx-qbserver
CURRENT VERSION: 1.0.0
FRAMEWORK: QBCore
PRICE: Free**

I got bored so i made a server, Didn't know what to do with it. So i released it here for anyone to use it.

smdx-qbserver is a nearly fully built server that has been worked on for about a month. This server is built in the way we construct our servers. We use a variety of scripts from several prominent creators, including Jimathy, Project Sloth, and Mustache Dom's scripts, among others! This server should be completely bug-free, but don't be surprised if you encounter a bug or two. (If you do, please report them to us, and we will fix them.)

We also give great credit to all the creators whose scripts we have used. (See the readme files for details.)

**Feel free to give us feedback of it!**